@@ -51,17 +51,13 @@ public:
     using A::A;
 
     // allocate more than caller required
-    T* allocate(size_t n) {
+    constexpr T* allocate(size_t n) {
         T* x = A::allocate(n + RawAllocator::_trailing);
-        return x;
-    }
-    T* allocate(size_t n, const void* hint) {
-        T* x = A::allocate(n + RawAllocator::_trailing, hint);
         return x;
     }
 
     // deallocate the storage referenced by the pointer p
-    void deallocate(T* p, size_t n) { A::deallocate(p, n + RawAllocator::_trailing); }
+    constexpr void deallocate(T* p, size_t n) { A::deallocate(p, (n + RawAllocator::_trailing)); }
 
     // do not initialized allocated.
     template <typename U>
@@ -165,8 +161,14 @@ inline void make_room(std::vector<T>* v, size_t n) {
     v->swap(reinterpret_cast<std::vector<T>&>(rv));
 }
 
-inline void make_room(std::string* s, size_t n) {
+inline void make_room_pad16(std::string* s, size_t n) {
     RawStringPad16 rs;
+    rs.resize(n);
+    s->swap(reinterpret_cast<std::string&>(rs));
+}
+
+inline void make_room(std::string* s, size_t n) {
+    RawString rs;
     rs.resize(n);
     s->swap(reinterpret_cast<std::string&>(rs));
 }
